canli_satis_mvp/server/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.6
sqlalchemy==2.0.32
psycopg[binary]==3.2.1
python-jose==3.3.0
passlib[bcrypt]==1.7.4
pydantic==2.8.2
reportlab==4.2.2
apscheduler==3.10.4
python-dotenv==1.0.1
